requirements.txt
django
gunicorn
django-heroku
django-localflavor
Pillow
whitenoise
